+++ plugins/jsx.ts
@@ -3,7 +3,14 @@ import loader from "../core/loaders/module.ts";
 import { merge, parseJSX } from "../core/utils.ts";
 import { dirname, join, toFileUrl } from "../deps/path.ts";
 
-import type { Data, Engine, Helper, Site } from "../core.ts";
+import type {
+  Data,
+  DenoConfig,
+  Engine,
+  Helper,
+  ImportMap,
+  Site,
+} from "../core.ts";
 
 export interface Options {
   /** The list of extensions this plugin applies to */
@@ -89,6 +96,16 @@ export class JsxEngine implements Engine {
   addHelper(name: string, fn: Helper) {
     this.helpers[name] = fn;
   }
+}
+
+/** Configure this plugin on "lume init" */
+export function init(importMap: ImportMap, denoConfig: DenoConfig) {
+  importMap.imports["react/jsx-runtime"] = import.meta.resolve(
+    "../deps/react.ts",
+  );
+  denoConfig.compilerOptions ||= {};
+  denoConfig.compilerOptions.jsx = "react-jsx";
+  denoConfig.compilerOptions.jsxImportSource = "react";
 }
 
 /** Register the plugin to support JSX and TSX files */