+++ plugins/jsx_preact.ts
@@ -8,7 +8,14 @@ import loader from "../core/loaders/module.ts";
 import { merge, parseJSX } from "../core/utils.ts";
 import { dirname, join, toFileUrl } from "../deps/path.ts";
 
-import type { Data, Engine, Helper, Site } from "../core.ts";
+import type {
+  Data,
+  DenoConfig,
+  Engine,
+  Helper,
+  ImportMap,
+  Site,
+} from "../core.ts";
 import type { ComponentChildren } from "../deps/preact.ts";
 
 export interface Options {
@@ -93,6 +100,16 @@ export class PreactJsxEngine implements Engine {
   addHelper(name: string, fn: Helper) {
     this.helpers[name] = fn;
   }
+}
+
+/** Configure this plugin on "lume init" */
+export function init(importMap: ImportMap, denoConfig: DenoConfig) {
+  importMap.imports["preact/jsx-runtime"] = import.meta.resolve(
+    "../deps/preact.ts",
+  );
+  denoConfig.compilerOptions ||= {};
+  denoConfig.compilerOptions.jsx = "react-jsx";
+  denoConfig.compilerOptions.jsxImportSource = "preact";
 }
 
 /** Register the plugin to support JSX and TSX files */